+++ Particle.h
@@ -13,90 +13,65 @@
 
 #define DECAY_RATE 0.001f
 
-class Particle {
-private:
-	Vector2f position;
-	Vector2f velocity;
-	float lifespan;
-	float current_lifespan;
+namespace pe {
+	class Particle {
+	private:
+		Vector2f position;
+		Vector2f velocity;
+		float lifespan;
+		float current_lifespan;
 
-public:
-	/* Constructors& Deconstructors */ 
-	Particle(const Vector2f&, const Vector2f&, const float);
-	~Particle() {};
+	public:
+		/* Constructors& Deconstructors */
+		Particle(const Vector2f&, const Vector2f&, const float);
+		~Particle() {};
 
-	/* Game Loop Functions */
-	/// <summary>
-	/// Update values of a particle (i.e. position, lifespan)
-	/// </summary>
-	void update();
+		/* Game Loop Functions */
+		/// <summary>
+		/// Update values of a particle (i.e. position, lifespan)
+		/// </summary>
+		void update();
 
-	/// <summary>
-	/// Check to make sure particle is not null.
-	/// Render particle and draw to window.
-	/// </summary>
-	/// <param name=""></param>
-	virtual void render(RenderWindow&) = 0;
+		/// <summary>
+		/// Check to make sure particle is not null.
+		/// Render particle and draw to window.
+		/// </summary>
+		/// <param name=""></param>
+		virtual void render(RenderWindow&) = 0;
 
-	/* Accessors& Mutators */
+		/* Accessors& Mutators */
 
-	/// <summary>
-	/// Return the position of the particle as a sf::Vector2f
-	/// </summary>
-	Vector2f& getPosition();
+		/// <summary>
+		/// Return the position of the particle as a sf::Vector2f
+		/// </summary>
+		Vector2f& getPosition();
 
-	Vector2f& getVelocity() const;
+		Vector2f& getVelocity();
 
-	float getLifespan() const;
+		float getLifespan() const;
 
-	float getLifespanRemaining() const;
+		float getLifespanRemaining() const;
 
-	void setPosition(const Vector2f&);
+		void setPosition(const Vector2f&);
 
-	void setVelocity(const Vector2f&);
+		void setVelocity(const Vector2f&);
 
-	void setLifeSpan(const float);
+		void setLifeSpan(const float);
 
-	/* Helper Functions */
+		/* Helper Functions */
 
-	/// <summary>
-	/// Returns true if the lifespan is less than or equal to 0.
-	/// Else return false.
-	/// </summary>
-	bool isAlive() const;
+		/// <summary>
+		/// Returns true if the lifespan is less than or equal to 0.
+		/// Else return false.
+		/// </summary>
+		bool isAlive() const;
 
-	/// <summary>
-	/// Set the lifespan to a stored float value
-	/// </summary>
-	/// <param name=""></param>
-	void setAlive();
-};
-
-class ShapeParticle : protected Particle {
-private:
-	Shape* shape;
-
-public:
-	ShapeParticle(const Vector2f& iP, const Vector2f& iV, const float ls) :
-		Particle(iP, iV, ls) {};
-	~ShapeParticle();	// needs to delete contents of shape if dynamically 
-						// allocated
-	void update() {};
-	virtual void render(RenderWindow&) = 0;
-	inline Shape* GetShape() { return shape; }
-};
-
-class CircleParticle : public ShapeParticle {
-private:
-	float radius;
-
-public:
-	CircleParticle(const float);
-	~CircleParticle() {};
-	void update() {};
-	virtual void render(RenderWindow&);
-	float GetRadius() const;
-	void setRadius(const float);
-};
+		/// <summary>
+		/// Set the lifespan to a stored float value
+		/// </summary>
+		/// <param name=""></param>
+		void setAlive();
+	};
+}
 
 #endif // !PARTICLE_H
